End of Section 5

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -9,6 +9,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -54,12 +55,13 @@ namespace AuctionService.Controllers
             return _mapper.Map<AuctionDto>(auction);
         }
 
+        [Authorize]
         [HttpPost]
         public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto auctionDto)
         {
             var auction = _mapper.Map<Auction>(auctionDto);
-            // TODO: Add current user as seller
-            auction.Seller = "test";
+            
+            auction.Seller = User.Identity.Name;
             _context.Auctions.Add(auction);
 
             var newAuction =  _mapper.Map<AuctionDto>(auction);
@@ -74,6 +76,7 @@ namespace AuctionService.Controllers
                 new {auction.Id}, newAuction);
         }
 
+        [Authorize]
         [HttpPut("{id}")]
         public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDto updateAuctionDto)
         {
@@ -82,7 +85,7 @@ namespace AuctionService.Controllers
 
             if(auction == null) return NotFound();
 
-            // TODO: Check seller name = user name
+            if(auction.Seller != User.Identity.Name) return Forbid();  // http:403 response
 
             auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
             auction.Item.Model = updateAuctionDto.Model?? auction.Item.Model;
@@ -97,13 +100,14 @@ namespace AuctionService.Controllers
             return BadRequest("Problem saving changes");
         }
 
+        [Authorize]
         [HttpDelete("{id}")]
         public async Task<ActionResult> DeleteAuction(Guid id)
         {
             var auction = await _context.Auctions.FindAsync(id);
             if(auction == null) return NotFound();
 
-            // TODO:  Check seller is user name
+            if(auction.Seller != User.Identity.Name) return Forbid();
             _context.Auctions.Remove(auction);
 
             await _publishedEndPoint.Publish<AuctionDeleted>(new {Id = auction.Id.ToString() });
